# docs/api-spec.yaml
openapi: 3.0.3
info:
  title: VertexLearn AI API
  version: 0.7.0
  description: Phase 7 surface — auth, course foundation, learning core, AI tutor/RAG (course-isolated), summaries, AI quiz drafts, flashcards, study plans, mastery, recommendations, admin workspace, course approvals, discussions, announcements, notifications, email abstraction, S3/MinIO storage (uploads, signed URLs), catalog filters/reviews, payments/revenue, Redis caching (catalog/course/leaderboard, x-cache), rate limits (auth 10/min/IP, AI 20/min/user, global 100/min), gamification leaderboard, health/ready aliases.
servers:
  - url: http://localhost:4000
paths:
...
  /api/v1/admin/users:
    get:
      summary: List/search users (admin only; supports q + role + pagination)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: page, in: query, schema: { type: integer, default: 1 } }
        - { name: pageSize, in: query, schema: { type: integer, default: 20 } }
        - { name: q, in: query, schema: { type: string } }
        - { name: role, in: query, schema: { type: string, enum: [student, instructor, admin] } }
      responses:
        '200': { description: Paginated users with roles }
        '401': { description: Unauthorized }
        '403': { description: Admin only }
  /api/v1/admin/users/{id}/role:
    put:
      summary: Assign/revoke a role (admin only; cannot revoke own admin; cannot revoke last admin)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Updated roles }
        '403': { description: Admin only or self-revoke }
        '404': { description: User not found }
        '409': { description: Last admin }
  /api/v1/admin/users/{id}/suspend:
    put:
      summary: Suspend/restore a user (admin only; suspends revoke refresh tokens and block login)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Suspension updated }
        '403': { description: Admin only or self-suspend }
        '404': { description: User not found }
  /api/v1/admin/users/{id}:
    delete:
      summary: Delete a user (admin only; blocked for self, last admin, and course owners)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '204': { description: Deleted }
        '403': { description: Admin only or self-delete }
        '404': { description: User not found }
        '409': { description: Last admin or owns courses }
  /api/v1/admin/courses/pending:
    get:
      summary: Pending course approval queue (admin only; uses the existing course status model)
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Pending courses }
        '403': { description: Admin only }
  /api/v1/admin/courses/{id}/approve:
    post:
      summary: Approve a course (pending → published) with approval record + instructor notification
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Approved course }
        '403': { description: Admin only }
        '404': { description: Course not found }
  /api/v1/admin/courses/{id}/reject:
    post:
      summary: Reject a course (pending → rejected) with approval record + instructor notification
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Rejected course }
        '403': { description: Admin only }
        '404': { description: Course not found }
  /api/v1/admin/courses/{id}/decision:
    post:
      summary: 'Unified approve/reject decision ({decision: approved|rejected, comment})'
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Decided course }
        '403': { description: Admin only }
        '404': { description: Course not found }
  /api/v1/admin/courses/{id}/approvals:
    get:
      summary: Approval history for a course (reviewer, decision, comment, timestamps)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: id, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Approval records }
        '404': { description: Course not found }
  /api/v1/admin/analytics/overview:
    get:
      summary: Platform analytics from real data (users, enrollments, completion rate, DAU estimate; revenue aggregates completed payments — zero when none recorded)
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Analytics overview }
        '403': { description: Admin only }
  /api/v1/admin/moderation/flagged-posts:
    get:
      summary: Flagged forum posts queue (admin, platform-wide)
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Flagged posts }
        '403': { description: Admin only }
  /api/v1/admin/moderation/posts/{postId}:
    put:
      summary: Moderate a flagged post (resolve|dismiss|hide|unhide|delete)
      security: [{ bearerAuth: [] }]
      parameters:
        - { name: postId, in: path, required: true, schema: { type: string, format: uuid } }
      responses:
        '200': { description: Moderated }
        '403': { description: Admin only }
        '404': { description: Post not found }
...
  /api/v1/admin/payments:
    get:
      summary: List payment records (admin only)
      security: [{ bearerAuth: [] }]
      responses:
        '200': { description: Payments }
        '403': { description: Admin only }
    post:
      summary: Record a payment (admin only; real ingress for revenue analytics; no gateway)
      security: [{ bearerAuth: [] }]
      responses:
        '201': { description: Payment recorded }
        '403': { description: Admin only }
        '404': { description: User or course not found }
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT